feat: 💄UX/UI Improvements on HiddenCollections section
