+++ .changeset/short-hotels-invite.md
@@ -1,0 +1,5 @@
+---
+"ledger-live-desktop": patch
+---
+
+UX/UI Improvements on HiddenCollections section
