+++ apps/ledger-live-desktop/src/renderer/screens/settings/sections/Accounts/HiddenNFTCollections/helpers.ts
@@ -1,0 +1,2 @@
+export const splitAddress = (str: string, length: number) =>
+  `${str.slice(0, length)}...${str.slice(-length)}`;
